spark_stream/requirements.txt
pyspark
redis
kafka-python
